+++ Domain.Supply/Entities/MedicalSupply/MedicalSupplyEntity.cs
@@ -1,13 +1,12 @@
 ﻿using System.Collections.Generic;
-using Diagnosea.Submarine.Abstractions.Enums;
 using Diagnosea.Submarine.Abstractions.Enums.Supply;
 
-namespace Diagnosea.Submarine.Domain.Supply.Entities.Supply
+namespace Diagnosea.Submarine.Domain.Supply.Entities.MedicalSupply
 {
     public class MedicalSupplyEntity : SupplyEntity
     {
-        public IList<Ailment> Treats { get; set; }
-
+        public IList<MedicalSupplyTreatmentEntity> Treats { get; set; }
+        
         public MedicalSupplyEntity()
         {
             Category = SupplyCategory.Medical;
