+++ Domain.Supply/Entities/MedicalSupply/MedicalSupplyTreatmentEntity.cs
@@ -1,0 +1,12 @@
+﻿using Diagnosea.Submarine.Abstractions.Enums;
+
+namespace Diagnosea.Submarine.Domain.Supply.Entities.MedicalSupply
+{
+    public class MedicalSupplyTreatmentEntity
+    {
+        public Ailment Treating { get; set; }
+        public int Dosage { get; set; }
+        public Occurence Occurence { get; set; }
+        public int Period { get; set; }
+    }
+}
