+++ Domain.Supply/Entities/SupplyRequirementEntity.cs
@@ -1,13 +1,12 @@
 ﻿using System;
 using Diagnosea.Submarine.Abstractions.Enums;
 
-namespace Diagnosea.Submarine.Domain.Supply.Entities.Supply.SupplyCompatibility
+namespace Diagnosea.Submarine.Domain.Supply.Entities
 {
-    public class SupplyCompatibilityEntity
+    public class SupplyRequirementEntity
     {
         public Guid Id { get; set; }
         public Guid SupplyId { get; set; }
-        public string Description { get; set; }
         public int Measurement { get; set; }
         public Metric Metric { get; set; }
     }
