--- Domain.Supply/Entities/SupplyStore/SupplyStoreEntity.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Diagnosea.Submarine.Domain.Supply.Entities.SupplyStore
-{
-    public class SupplyStoreEntity
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public string Website { get; set; }
-    }
-}
